feat(BuySell): add buy sell coordinator and set as delegate
