+++ Blockchain/Wallet.m
@@ -683,10 +683,6 @@
         [weakSelf on_get_pending_trades_error:error];
     };
 
-    self.context[@"objc_initialize_webview"] = ^() {
-        [weakSelf initialize_webview];
-    };
-
 #pragma mark Settings
     
     self.context[@"objc_on_get_account_info_and_exchange_rates"] = ^() {
@@ -4045,15 +4041,6 @@
 - (void)on_get_pending_trades_error:(JSValue *)error
 {
     [[AlertViewPresenter sharedInstance] standardNotifyWithMessage:[error toString] title:BC_STRING_ERROR handler: nil];
-}
-
-- (void)initialize_webview
-{
-    if ([self.delegate respondsToSelector:@selector(initializeWebView)]) {
-        [self.delegate initializeWebView];
-    } else {
-        DLog(@"Error: delegate of class %@ does not respond to selector initializeWebView!", [delegate class]);
-    }
 }
 
 - (void)on_fetch_eth_history_success
