+++ Blockchain/js/wallet-ios.js
@@ -2325,12 +2325,6 @@ MyWalletPhone.isBuyFeatureEnabled = function () {
   return userHasAccess && wallet.external && canBuy(wallet.accountInfo, options)
 }
 
-MyWalletPhone.setupBuySellWebview = function() {
-    walletOptions.fetch().then(function() {
-        objc_initialize_webview();
-    });
-}
-
 MyWalletPhone.getBuySellWebviewRootURL = function() {
     var options = walletOptions.getValue();
     var mobile = options.mobile;
